backend/requirements.txt
asgiref==3.8.1
Django==5.1.5
sqlparse==0.5.3
psycopg2-binary==2.9.10
djangorestframework==3.16.0
drf_yasg==1.21.10
requests==2.32.3
django-cors-headers==4.7.0
